create general components
